device has sensors, room has device
